requirements.txt
cachetools==4.2.4
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.0.4
dialogflow==2.0.0
Flask==2.0.3
google-api-core==1.31.5
google-auth==1.35.0
google-cloud-dialogflow==2.12.0
googleapis-common-protos==1.55.0
grpcio==1.44.0
gunicorn==20.1.0
idna==3.3
itsdangerous==2.1.0
Jinja2==3.0.3
MarkupSafe==2.1.0
packaging==21.3
proto-plus==1.20.3
protobuf==3.19.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyJWT==2.3.0
pyparsing==3.0.7
pytz==2021.3
requests==2.27.1
rsa==4.8
six==1.16.0
twilio==7.7.0
urllib3==1.26.8
Werkzeug==2.0.3
